Replace code with working one
No need to wrap tryParse in try as long as the value provided is not null, which we've covered in the first clause. Moreover, code doesn't actually check whether a number is provided - int.tryParse(oib.substring(i, i + 1)) ?? 0 returns 0 if value is not numerical. Function name was in English and variables were in Croatian, among other issues with their names.